commit from git

+++ DKPlugins/DKEventTargetTest/DKPointerEvent.h
@@ -39,9 +39,9 @@ public:
 	// [PointerEvent.height](Read only) https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent/height
 	unsigned int height;
 	// [PointerEvent.pressure](Read only) https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent/pressure
-	float pressure;
+	double pressure;
 	// [PointerEvent.tangentialPressure) https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent/tangentialPressure
-	float tangentialPressure;
+	double tangentialPressure;
 	// [PointerEvent.tiltX](Read only) https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent/tiltX
 	int tiltX;
 	// [PointerEvent.tiltY](Read only) https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent/tiltY
